Remove dep, move to common component folder
