+++ web/package.json
@@ -39,6 +39,7 @@
     "axios": "^0.21.1",
     "clsx": "^1.1.0",
     "csv-string": "^4.0.1",
+    "date-fns": "^2.28.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "esri-leaflet": "3.0.3",
@@ -46,7 +47,6 @@
     "leaflet": "^1.7.1",
     "lodash": "^4.17.21",
     "luxon": "^2.1.1",
-    "materialui-daterange-picker": "^1.1.92",
     "ol": "^6.5.0",
     "plotly.js": "^2.5.1",
     "prettier": "^2.4.1",
