+++ web/src/features/fwiCalculator/components/FWIDateRange.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
-import { DateRange, DateRangePicker } from 'materialui-daterange-picker'
 import { Dialog } from '@material-ui/core'
 import { isUndefined } from 'lodash'
 import { DateTime } from 'luxon'
 import { PST_UTC_OFFSET } from 'utils/constants'
+import DateRangePickerWrapperMod from 'components/dateRangePicker/DateRangePickerWrapperMod'
+import { DateRange } from 'components/dateRangePicker/types'
 
 export interface FWIDateRangeProps {
   open: boolean
@@ -54,7 +55,7 @@ const FWIDateRange = ({
 
   return (
     <Dialog open={open} onClose={toggle}>
-      <DateRangePicker
+      <DateRangePickerWrapperMod
         initialDateRange={{ startDate, endDate }}
         open={open}
         toggle={toggle}
