+++ web/src/features/hfiCalculator/components/HFIPageSubHeader.tsx
@@ -7,10 +7,10 @@ import { isUndefined } from 'lodash'
 import { FireCentre } from 'api/hfiCalcAPI'
 import AboutDataModal from 'features/hfiCalculator/components/AboutDataModal'
 import { HelpOutlineOutlined } from '@material-ui/icons'
-import { DateRange } from 'materialui-daterange-picker'
 import { formControlStyles } from 'app/theme'
 import LastUpdatedHeader from 'features/hfiCalculator/components/LastUpdatedHeader'
 import { HFIResultResponse } from 'features/hfiCalculator/slices/hfiCalculatorSlice'
+import { DateRange } from 'components/dateRangePicker/types'
 import PrepDateRangeSelector from 'features/hfiCalculator/components/PrepDateRangeSelector'
 
 const useStyles = makeStyles(theme => ({
